rename lib to vm_url
