+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests_with_privacy/test_ttest_onesample.py
@@ -12,7 +12,7 @@ import rpy2.robjects as robjects
 import sys
 from os import path
 sys.path.append(path.abspath(__file__))
-from tests.lib import vmUrl
+from tests.vm_url import vmUrl
 endpointUrl= vmUrl+'TTEST_ONESAMPLE'
 folderPath = 'R_scripts'
 file ='ttest_onesample.Rmd'
