+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests_without_privacy/test_CalibrationBelt.py
@@ -2,7 +2,7 @@ import pytest
 import json
 import requests
 import math
-from tests.lib import vmUrl
+from tests.vm_url import vmUrl
 endpointUrl= vmUrl+'CALIBRATION_BELT'
 
 def get_test_params():
