+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests_without_privacy/test_PearsonCorrelation.py
@@ -2,7 +2,7 @@ import requests
 import json
 import logging
 import math
-from tests.lib import vmUrl
+from tests.vm_url import vmUrl
 endpointUrl= vmUrl+'PEARSON_CORRELATION'
 
 # Required datasets: adni_9rows, adni, data_pr1, desd-synthdata
