+++ Exareme-Docker/src/mip-algorithms/tests/exareme_tests/test_wrongDataset.py
@@ -4,7 +4,7 @@ import json
 import logging
 import sys
 sys.path.insert(0,'../')
-from tests.algorithm_tests.lib import vmUrl
+from tests.vm_url import vmUrl
 
 endpointUrl=vmUrl+'LINEAR_REGRESSION'
